Before deleting an user, ask for confirmation.

+++ resources/js/views/users/Users.vue
@@ -8,17 +8,17 @@
             <template v-slot:cell(actions)="data">
                 <b-button variant="primary" size="sm" title="Edit" @click="data.toggleDetails"
                           :disabled="disabled || isDeletingId === data.item.id">
-                    <fa-icon icon="edit" />
+                    <fa-icon icon="edit"/>
                 </b-button>
                 <b-button v-if="user.id !== data.item.id" variant="danger" size="sm" title="Delete"
                           @click="handleDelete(data.item)" :disabled="disabled || isDeletingId === data.item.id">
-                    <fa-icon v-if="isDeletingId === data.item.id" icon="spinner" spin />
-                    <fa-icon v-else icon="trash" />
+                    <fa-icon v-if="isDeletingId === data.item.id" icon="spinner" spin/>
+                    <fa-icon v-else icon="trash"/>
                 </b-button>
             </template>
             <template v-slot:row-details="row">
                 <b-card>
-                    <user-form :user="row.item" @cancel="row.toggleDetails" />
+                    <user-form :user="row.item" @cancel="row.toggleDetails"/>
                 </b-card>
             </template>
         </b-table>
@@ -27,14 +27,15 @@
 
 <script>
     import _ from 'lodash';
-    import { mapState } from 'vuex';
+    import {mapState} from 'vuex';
     import {DELETE_USER, LOAD_USERS} from '../../store/actions';
     import UserForm from './UserForm.vue';
     import ToastMixin from '../../mixins/ToastMixin';
+    import ConfirmModalMixin from '../../mixins/ConfirmModalMixin';
 
     export default {
         name: 'Users',
-        mixins: [ToastMixin],
+        mixins: [ToastMixin, ConfirmModalMixin],
         components: {
             UserForm
         },
@@ -95,11 +96,17 @@
                     .finally(() => this.isLoading = false);
             },
             handleDelete(user) {
-                this.isDeletingId = user.id;
-                return this.$store.dispatch(DELETE_USER, user)
-                    .then(() => this.showSuccess('User Deleted'))
-                    .catch(error => this.showNetworkError(error))
-                    .finally(() => this.isDeletingId = null);
+                return this.showConfirmDeletionModal('Are you sure to delete this user?')
+                    .then(value => {
+                        if (value) {
+                            this.isDeletingId = user.id;
+                            return this.$store.dispatch(DELETE_USER, user)
+                                .then(() => this.showSuccess('User Deleted'))
+                                .catch(error => this.showNetworkError(error))
+                                .finally(() => this.isDeletingId = null);
+                        }
+                    });
+
             }
         }
     }
